requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
amqp==5.2.0
    # via kombu
asgiref==3.7.2
    # via django
billiard==4.2.0
    # via celery
celery==5.3.5
    # via -r requirements.in
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
cryptography==41.0.5
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.2.7
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-ckeditor
    #   django-js-asset
django-allauth==0.58.2
    # via -r requirements.in
django-ckeditor==6.7.0
    # via -r requirements.in
django-js-asset==2.1.0
    # via django-ckeditor
idna==3.4
    # via requests
kombu==5.3.3
    # via celery
oauthlib==3.2.2
    # via requests-oauthlib
pillow==10.1.0
    # via -r requirements.in
prompt-toolkit==3.0.40
    # via click-repl
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   django-allauth
    #   pyjwt
python-dateutil==2.8.2
    # via celery
python-dotenv==1.0.0
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via
    #   celery
    #   django
urllib3==2.0.7
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.9
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
